+++ tests/test-ttyname_r.c
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2023 Gaël PORTAY
+ *
+ * SPDX-License-Identifier: LGPL-2.1-or-later
+ */
+
+#include <stdlib.h>
+#include <stdio.h>
+#include <limits.h>
+
+#include <unistd.h>
+
+int main(int argc, char * const argv[])
+{
+	char buf[PATH_MAX];
+	int fd = -1;
+
+	if (argc < 2) {
+		fprintf(stderr, "Too few arguments\n");
+		exit(EXIT_FAILURE);
+	} else if (argc > 2) {
+		fprintf(stderr, "Too many arguments\n");
+		exit(EXIT_FAILURE);
+	}
+
+	fd = strtoul(argv[1], NULL, 0);
+
+	if (ttyname_r(fd, buf, sizeof(buf)) == -1) {
+		perror("ttyname_r");
+		return EXIT_FAILURE;
+	}
+
+	printf("%s\n", buf);
+
+	return EXIT_SUCCESS;
+}