+++ ttyname_r.c
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2023 Gaël PORTAY
+ *
+ * SPDX-License-Identifier: LGPL-2.1-or-later
+ */
+
+#include <stdio.h>
+#include <errno.h>
+#include <dlfcn.h>
+
+#include <unistd.h>
+
+#include "iamroot.h"
+
+__attribute__((visibility("hidden")))
+int next_ttyname_r(int fd, char *buf, size_t bufsize)
+{
+	int (*sym)(int, char *, size_t);
+	int ret;
+
+	sym = dlsym(RTLD_NEXT, "ttyname_r");
+	if (!sym)
+		return __set_errno(ENOSYS, -1);
+
+	ret = sym(fd, buf, bufsize);
+	if (ret == -1)
+		__fpathperror(fd, __func__);
+
+	return ret;
+}
+
+int ttyname_r(int fd, char *buf, size_t bufsize)
+{
+	char *s;
+	int ret;
+
+	ret = next_ttyname_r(fd, buf, bufsize);
+	if (ret == -1)
+		return -1;
+
+	s = __striprootdir(buf);
+	if (!s) {
+		__fpathperror(fd, __func__);
+		return -1;
+	}
+
+	__debug("%s(fd: %i <-> '%s', ...)\n", __func__, fd, __fpath(fd));
+
+	return ret;
+}